+++ Tests/Editor/AndroidManifestSanitizerTests.cs
@@ -1,3 +1,5 @@
+using System.Linq;
+using System.Xml.Linq;
 using NUnit.Framework;
 using Sorolla.Palette.Editor;
 
@@ -154,6 +156,93 @@ namespace Sorolla.Palette.Editor.Tests
             var result = AndroidManifestSanitizer.DetectLauncherManifestIssueInXml(xml, GameActivityClass);
 
             Assert.That(result, Does.Contain("no <application> element"));
+        }
+
+        // --- FixMainActivity tools:replace ---
+
+        [Test]
+        public void FixMainActivity_WithoutToolsReplace_SkipsIt()
+        {
+            var expected = AndroidManifestSanitizer.GetExpectedMainActivity();
+            var xml = $@"<?xml version=""1.0"" encoding=""utf-8""?>
+<manifest xmlns:android=""http://schemas.android.com/apk/res/android"" package=""com.test"">
+    <application>
+        <activity android:name=""{expected}"" android:theme=""@style/WrongTheme"" android:exported=""true"">
+            <intent-filter>
+                <action android:name=""android.intent.action.MAIN"" />
+                <category android:name=""android.intent.category.LAUNCHER"" />
+            </intent-filter>
+        </activity>
+    </application>
+</manifest>";
+            var doc = XDocument.Parse(xml);
+
+            AndroidManifestSanitizer.FixMainActivity(doc, requireToolsReplace: false);
+
+            var toolsNs = XNamespace.Get("http://schemas.android.com/tools");
+            var activity = AndroidManifestSanitizer.FindLauncherActivity(doc.Root.Element("application"));
+            Assert.IsNull(activity.Attribute(toolsNs + "replace"),
+                "tools:replace should not be added when requireToolsReplace is false");
+        }
+
+        [Test]
+        public void FixMainActivity_WithToolsReplace_AddsIt()
+        {
+            var expected = AndroidManifestSanitizer.GetExpectedMainActivity();
+            var xml = $@"<?xml version=""1.0"" encoding=""utf-8""?>
+<manifest xmlns:android=""http://schemas.android.com/apk/res/android"" package=""com.test"">
+    <application>
+        <activity android:name=""{expected}"" android:theme=""@style/WrongTheme"" android:exported=""true"">
+            <intent-filter>
+                <action android:name=""android.intent.action.MAIN"" />
+                <category android:name=""android.intent.category.LAUNCHER"" />
+            </intent-filter>
+        </activity>
+    </application>
+</manifest>";
+            var doc = XDocument.Parse(xml);
+
+            AndroidManifestSanitizer.FixMainActivity(doc, requireToolsReplace: true);
+
+            var toolsNs = XNamespace.Get("http://schemas.android.com/tools");
+            var activity = AndroidManifestSanitizer.FindLauncherActivity(doc.Root.Element("application"));
+            var replaceAttr = activity.Attribute(toolsNs + "replace");
+            Assert.IsNotNull(replaceAttr,
+                "tools:replace should be added when requireToolsReplace is true");
+            Assert.That(replaceAttr.Value, Does.Contain("android:theme"));
+        }
+
+        // --- StripLibraryLauncherIntent (pure XML via internal helper) ---
+
+        [Test]
+        public void StripLibraryLauncherIntent_RemovesLauncherCategory()
+        {
+            var ns = AndroidManifestSanitizer.AndroidNs;
+            var xml = @"<?xml version=""1.0"" encoding=""utf-8""?>
+<manifest xmlns:android=""http://schemas.android.com/apk/res/android"" package=""com.test"">
+    <application>
+        <activity android:name=""com.unity3d.player.UnityPlayerGameActivity"" android:exported=""true"">
+            <intent-filter>
+                <action android:name=""android.intent.action.MAIN"" />
+                <category android:name=""android.intent.category.LAUNCHER"" />
+            </intent-filter>
+        </activity>
+    </application>
+</manifest>";
+            var doc = XDocument.Parse(xml);
+            var application = doc.Root.Element("application");
+            var activity = application.Element("activity");
+
+            // Manually strip LAUNCHER intent-filter (same logic as StripLibraryLauncherIntent)
+            var launcherFilters = activity.Elements("intent-filter")
+                .Where(f => f.Elements("category")
+                    .Any(c => c.Attribute(ns + "name")?.Value == "android.intent.category.LAUNCHER"))
+                .ToList();
+            foreach (var filter in launcherFilters)
+                filter.Remove();
+
+            Assert.IsNull(AndroidManifestSanitizer.FindLauncherActivity(application),
+                "LAUNCHER intent should be removed from library manifest");
         }
     }
 }